Extract generator download ans cleanup to generatorResolver.js
