+++ src/generate.js
@@ -10,6 +10,7 @@ const fetch = require("node-fetch");
 const shell = require("shelljs");
 const { parse } = require("yaml");
 
+const generatorResolver = require("./generatorResolver");
 const helpers = require("./helpers");
 const log = require("./log");
 const transformers = require("./transformers");
@@ -95,8 +96,6 @@ async function generate(schemaPathOrUrl, generatorPathOrUrl, options) {
   log.verbose("  \\___/ | .__/ \\___||_||_|/_/ \\_\\ |_|  |___|  |_|  \\___/|_|  \\__, | \\___|  ");
   log.verbose("        |_|                                                  |___/         ");
   log.verbose("");
-  let temporaryFolder;
-  let npmPackage = false;
   let exception = null;
   let numberOfDiscoveredModels = 0;
   let numberOfDiscoveredEndpoints = 0;
@@ -104,37 +103,13 @@ async function generate(schemaPathOrUrl, generatorPathOrUrl, options) {
     log.standard(`Loading generator from '${generatorPathOrUrl}'`);
     let generatorPath;
     if (isUrl(generatorPathOrUrl)) {
-      // if the generator is specified as a git URL, clone it into a temporary directory
-      if (generatorPathOrUrl.endsWith(".git")) {
-        generatorPath = temporaryFolder = fs.mkdtempSync(
-          path.join(os.tmpdir(), "generator")
-        );
-        log.verbose(`Cloning generator from ${generatorPathOrUrl} to ${generatorPath}`);
-        shell.exec(`git clone ${generatorPathOrUrl} ${generatorPath}`, {silent:true});
-      } else {
-        throw new Error(
-          `Generator URL ${generatorPathOrUrl} does not end with ".git", check that the URL points to a valid generator`
-        );
-      }
+      generatorPath = generatorResolver.cloneGenerator(generatorPathOrUrl);
     } else {
       //first check if there is a local generator
       generatorPath = path.resolve(generatorPathOrUrl);
       if (!fs.existsSync(generatorPath)) {
-        //if no local generator, assume it is npm package name.
-        log.verbose(`Checking if npm package ${generatorPathOrUrl} is installed`);
-        const currentPath = process.cwd();
-        shell.cd(__dirname, {silent:true});
-        if (!shell.exec(`npm list --depth=0`, {silent:true}).stdout.match(new RegExp(`^\\+--.${generatorPathOrUrl}@\\d+\.\\d+\.\\d+$`, 'm'))) {
-          npmPackage = true;
-          log.verbose(`npm package ${generatorPathOrUrl} doesn't exist, installing package`);
-          if (shell.exec(`npm install ${generatorPathOrUrl}`, {silent:true}).code !== 0) {
-            throw new Error(
-              `No local generator or npm package found using '${generatorPathOrUrl}', check that it points to a local generator or npm package`
-            );
-          }
-        }
-        generatorPath = path.resolve(`..\\node_modules\\${generatorPathOrUrl}`);
-        shell.cd(currentPath, {silent:true});
+        //if no local generator, assume it is an npm package name.
+        generatorPath = generatorResolver.installGeneratorFromNPM(generatorPathOrUrl);
       }
     }
 
@@ -246,17 +221,7 @@ async function generate(schemaPathOrUrl, generatorPathOrUrl, options) {
   } catch(e) {
     exception = e;
   } finally {
-    if (temporaryFolder) {
-      log.verbose(`Removing temporary folder ${temporaryFolder}`);
-      fs.rmSync(temporaryFolder, { recursive: true });
-    }
-    if(npmPackage) {
-      const currentPath = process.cwd();
-      shell.cd(__dirname, {silent:true});
-      log.verbose(`Removing npm package ${generatorPathOrUrl}`);
-      shell.exec(`npm uninstall ${generatorPathOrUrl}`, {silent:true});
-      shell.cd(currentPath, {silent:true});
-    }
+    generatorResolver.cleanup();
   }
   if (exception === null) {
     log.standard(`${log.divider}`);
